Add facilitator file
