+++ x402/Cargo.toml
@@ -7,9 +7,6 @@ edition = "2024"
 alloy.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
-axum.workspace = true
-chrono.workspace = true
-hex.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
