--- x402/src/axum.rs
@@ -1,23 +1,0 @@
-/// POST /verify
-/// Verifies a payment authorization without executing the transaction on the blockchain.
-pub async fn verify() -> Response {
-    todo!()
-}
-
-/// POST /settle
-/// Executes a verified payment by broadcasting the transaction to the blockchain.
-pub async fn settle() -> Response {
-    todo!()
-}
-
-/// GET /supported
-/// Returns the list of payment schemes and networks supported by the facilitator.
-pub async fn supported() -> Response {
-    todo!()
-}
-
-/// GET /discovery/resources
-pub async fn discovery() -> Response {
-    todo!()
-}
-
